==> Scripts/4.diff.abundance.mci.hc.Rmd <==
---
title: "Differential abundance Status MCI vs HC"
author: "Alba Troci"
date: "`r Sys.Date()`"  
output:
   #BiocStyle::html_document:
   html_document:
     
     code_folding: hide
editor_options:
  chunk_output_type: console
params:
  FIGPATH: "figures/"
  d.out: "./"
---

# Preparations
## Set global options

```{r style, results="asis", cache=FALSE, message = F, echo=FALSE}
# Set knit global options
library("knitr")
options(digits = 2, width = 80)
golden_ratio <- (1 + sqrt(5)) / 2
opts_chunk$set(echo = TRUE,
               tidy = FALSE,
               include = TRUE,
               fig.path = params$FIGPATH,
               dev=c("png",'pdf'),
               fig.height = 5,
               fig.width = 4 * golden_ratio,
               comment = '  ',
               dpi = 300,
               cache = FALSE)
  

# Pretty outputs
library("rmarkdown")
library("BiocStyle")
library("ggthemes") # add ggplot2 themes
library("ggplot2")
library("ggpubr")

# Set seed for reproducibility
set.seed(100)

# Set plotting theme
theme_set(theme_few(base_size = 14))

# Set output directory
d.out <- params$d.out
rm(params)
```


## Load libraries for the session
```{r libraries, message=FALSE}
library("magrittr")
library("ggplot2")
library("tidyverse")
library("readxl")
library("dada2")
library("phyloseq")
library("gridExtra")
library("vegan")
library("plyr")
library("scales")
library("reshape2")
library("DESeq2")
library("dplyr") #for data manipulation
library("msa") #multiple sequence alignment
library("ape") #to create phylogenetic tree
library("randomForest") 
library("caret") 
library("broom")
library("mlbench")
library("plotROC")
library("gg3D")
library("scatterplot3d")
library("car")
library("rgl")
library("rglwidget")

```

# Import files

Phyloseq object AD filtered samples
I have filtered out ASV with prevalence less than 5% in all samples
```{r}
# Phyloseq object AD filtered samples
d.ps.filter.ad.st <- "/home/alba/Desktop/Alba/Practice/AD/AD.update.jan.2022/4.filter"
ps.filter.ad.st <- "/ps.filter.ad.stool.prevalence.rds" %>%
  paste0(d.ps.filter.ad.st,"/", .) %>% 
  readRDS() 
ps.filter.ad.st

ps.st <- ps.filter.ad.st

```

## Edit

Subset Group = c("Controls", "Disease", "Risiko")

```{r, results= 'hide'}
nrow(sample_data(ps.st)[sample_data(ps.st)$Group %in% "Controls"])
nrow(sample_data(ps.st)[sample_data(ps.st)$Group %in% "Disease"])
nrow(sample_data(ps.st)[sample_data(ps.st)$Group %in% "Risiko"])
```

```{r}
#add one column for category
sample_data(ps.st)$Group <- as.character(sample_data(ps.st)$Group)
sample_data(ps.st)$Status[sample_data(ps.st)$Group == "Disease" & sample_data(ps.st)$Diagnose == "AD"] <- "AD"
sample_data(ps.st)$Status[sample_data(ps.st)$Group == "Disease" & sample_data(ps.st)$Diagnose == "MCI"] <- "MCI"
sample_data(ps.st)$Status[sample_data(ps.st)$Group == "Risiko" ] <- "Risiko"
sample_data(ps.st)$Status[sample_data(ps.st)$Group == "Controls" ] <- "HC"
sample_data(ps.st)$Group[sample_data(ps.st)$Group == "Controls" ] <- "HC"

```

```{r}
ps.dc <- subset_samples(ps.st, Group %in% c("Disease", "HC", "Risiko"))
ps.dc

sam <- data.frame(sample_data(ps.dc))

```



# OTU differential abundance testing with DESeq2
**Aim:** define taxa which are resposible for the difference between different Groups.

**Method:** DESeq2.

**Algorithm process:**

  * Input a phyloseq non normalized object.
  * Normalization (estimate size factors with RLE method).
  * Evaluate disperssion and genewise disperssion.
  * Negative Binomial GLM fitting and Wald statistics.
  
Separate groups
I want to compare Disease (AD+MCI) vs HC

```{r}
ps1 <- subset_samples(ps.dc, Status %in% c("AD", "HC"))
ps3 <- subset_samples(ps.dc, Status %in% c("MCI", "HC"))
```

Convert phyloseq to deseq
```{r}
sample_data(ps3)$Alter <- as.factor(sample_data(ps3)$Alter)
sample_data(ps3)$Status <- as.factor(sample_data(ps3)$Status)
Group_DESeq <- phyloseq_to_deseq2(ps3, ~Alter+Status)
Group_DESeq

```

Calculate geometric means prior to estimate size factors
```{r}
Group_DESeq = estimateSizeFactors(Group_DESeq, type="poscounts")

```

We want a normalized count of at least 15 in 5 or more samples (https://support.bioconductor.org/p/65091/):
```{r}
#we want a normalized count of at least 15 in 4 or more samples:
nc <- counts(Group_DESeq, normalized=TRUE)
filter <-rowSums(nc >= 15) >= 5
Group_DESeq <- Group_DESeq[filter,]
Group_DESeq

#then run deseq
```

Run DESeq
```{r}
Group_DESeq <- estimateSizeFactors(Group_DESeq, type="poscounts")
Group_DESeq <- estimateDispersions(Group_DESeq,maxit=5000,fitType="local")
Group_DESeq <- nbinomWaldTest(Group_DESeq, maxit=5000)
```

Remove ASV that did not converge
```{r}
ddsClean <- Group_DESeq[which(mcols(Group_DESeq)$betaConv),]
ddsClean
```


## Investigate test results table

```{r}
res = results(ddsClean, contrast = c("Status", "MCI", "HC"))
```


```{r}
res = res[order(res$padj, na.last=NA), ]
alpha = 0.01
sigtab = res[(res$padj < alpha), ]
sigtab = cbind(as(sigtab, "data.frame"), 
               as(tax_table(ps1)[rownames(sigtab), ], "matrix"))
sigtab[1:5, 1:8]

```


## Plot Results

Nr of ASV with padj < 0.01
```{r}
sum(res$padj < 0.01, na.rm=TRUE )
```

MA plot
```{r, out.width="60%"}
maplot <- ggmaplot(res, fdr = 0.01, fc = 1, genenames = NULL,
                   detection_call = NULL, size = 1,
                   palette = c("#B31B21", "#1465AC", "darkgray"),
                   legend="top", top = 15,font.label = c("bold", 10),
                   label.rectangle = TRUE, font.legend = c("bold",12),
                   font.main = "bold")
maplot
```

Print unique phylum names
```{r}
#set manual colors to each phylum
unique(tax_table(ps3)[,"Phylum"])  
cols <- c("Acidobacteria"= "#61864B",
          "Actinobacteria" = "#C46226",
          "Bacteria (unc.)"= "#A96755",
          "Bacteroidetes" = "#666666" ,
          "Elusimicrobia"  = "#767F3E" ,
          "Firmicutes" = "#77942C",
          "Fusobacteria"= "#B1A80D",
          "Proteobacteria" = "#E1A703", 
          "Spirochaetes" = "#946A7C", 
          "Lentisphaerae"  ="#C48F10",
          "Synergistetes"= "#957130",
          "Candidatus_Saccharibacteria"="#7B6EA6",
          "Verrucomicrobia" = "#935DA8")           
```


Let's look at the ASVs that were significantly different between the two groups. 
Contrast: Disease vs HC (means that Disease has positive logFoldChange)

```{r, out.width="70%", warning=FALSE}
theme_set(theme_bw())
sigtabgen = subset(sigtab, !is.na(Genus))
# Phylum order
x = tapply(sigtabgen$log2FoldChange, sigtabgen$Phylum, function(x) max(x))
x = sort(x, TRUE)
sigtabgen$Phylum = factor(as.character(sigtabgen$Phylum), levels=names(x))
# Genus order
x = tapply(sigtabgen$log2FoldChange, sigtabgen$Genus, function(x) max(x))
x = sort(x, TRUE)
sigtabgen$Genus = factor(as.character(sigtabgen$Genus), levels=names(x))
fig <- ggplot(sigtabgen, aes(y=Genus, x=log2FoldChange, color=Phylum)) + 
  geom_vline(xintercept = 0.0, color = "gray", size = 0.5) +
  geom_point(aes(size=sigtabgen$baseMean)) + 
  scale_size(range = c(0.1,3))+
  labs( size = "baseMean")+
  ggtitle("      HC                                  MCI")+
  theme(axis.text.x = element_text(angle = -90, hjust = 0, vjust=0.5), 
        text = element_text(size = 10),
        plot.title = element_text(size = 7)) +
    scale_colour_manual(values = cols)+
    scale_fill_manual(values = cols)


fig

ggsave("mci.vs.hc.png")

```

```{r}
library("rio")

export(sigtabgen, "mci.vs.hc.xlsx")



```


# References
1. https://bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#group-specific-condition-effects-individuals-nested-within-groups

2. https://support.bioconductor.org/p/64480/

3. http://bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html


# Session info
